fix: queue and message composition
