+++ server/Makefile
@@ -21,6 +21,7 @@ SRC += ./src/error/check_teams_names.c
 SRC += ./src/error/check_freq.c
 SRC += ./src/error/check_port.c
 SRC += ./src/error/help_flag.c
+SRC += ./src/queue.c
 
 OBJ = $(SRC:.c=.o)
